split up all tests to match src/table
